requirements.txt
numpy==1.26.2
opencv-python==4.9.0.80
Pillow==10.1.0
pygame
